+++ RELEASENOTES.md
@@ -4,6 +4,7 @@ These are the release notes for the [mediawiki-datamodel](README.md).
 
 * Title::getTitle is deprecated
 * Title::getText introduced to replace getTitle
+* Implemented File class
 
 ## Version 0.5 (2015-01-13)